Check permissions when deleting items

+++ backend/src/resolvers/Mutation.js
@@ -46,9 +46,17 @@ const mutations = {
   async deleteItem(parent, args, context, info) {
     const where = { id: args.id };
     // 1 find the items
-    // const item = await context.db.query.item({ where }, `{id title}`);
+    const item = await context.db.query.item({ where }, `{id title user {id}}`);
     // 2 check if they own that item or have the permissions
     //TODO
+    const ownsItem = item.user.id === context.request.userId;
+    // sum checks if at least one is true
+    const hasPermissions = context.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You can't do that!");
+    }
     // 3 delete
     return context.db.mutation.deleteItem({ where }, info);
   },
